Prime factors configurable. Octave channel sharing optional.

+++ MidiConnection.py
@@ -17,6 +17,7 @@ class MidiConnection:
             virtual=True
         )
 
+
     def get_message(self):
         return self.__port_in.poll()
 
@@ -26,7 +27,7 @@ class MidiConnection:
             mido.Message(
                 "pitchwheel",
                 channel=channel,
-                pitch=round(bending * 8191)
+                pitch=int(round(bending * 8191))
             )
         )
         self.__port_out.send(
